requirements.txt
torch==1.2.0
torchvision==0.4.0
easydict
matplotlib
opencv-python
glob3
pillow
dill
dominate>=2.3.1 
scikit-image
tensorboard==1.14.0
tensorboardX==1.9
tqdm==4.32.1
urllib3==1.25.8
visdom==0.1.8.9

